+++ anyblok/_logging.py
@@ -121,6 +121,7 @@ def init_logger(level='info', mode='console',
     :param filename: Out put file
     :param socket: Socket or UnixSocket
     :param facility:
+    ;exception: Exception
     """
     if logger.hasHandlers():
         for h in logger.handlers: