+++ anyblok/bloks/anyblok_core/__init__.py
@@ -2,9 +2,44 @@ from anyblok.blok import Blok
 
 
 class AnyBlokCore(Blok):
-    """ AnyBlok Core
+    """
+    This blok is required by all anyblok application. This blok define the main
+    fonctionnality to install, update and uninstall blok. And also list the
+    known models, fields, columns and relationships:
 
-    This blok is the main blok"""
+    * Core model
+    * Field Type
+        - Function
+    * Column Types:
+        - DateTime
+        - Decimal
+        - Float
+        - Time
+        - BigInteger
+        - Boolean
+        - Date
+        - Integer
+        - Interval
+        - LargeBinary
+        - SmallInteger
+        - String
+        - Text
+        - uString
+        - uText
+        - Selection
+        - Json
+    * Relationship types
+        - One2One
+        - Many2One
+        - One2Many
+        - Many2Many
+    * System Models
+        - Blok
+        - Model
+        - Field
+        - Column
+        - Relationship
+    """
 
     version = '0.0.1'
     autoinstall = True