+++ doc/CHANGES.rst
@@ -1,16 +1,20 @@
 CHANGES LOG
 ===========
 
-0.1
----
+0.1.0 (Next Version)
+--------------------
 
 Main version of AnyBlok. You can with this version
 
 * Create your own application
 * Connect on a database
 * Define blok
+* Install, Update, Uninstall the blok
+* Define field types
 * Define Column types
 * Define Relationship types
+* Define Core
+* Define Mixin
 * Define Model (SQL or not)
 * Define SQL view
 * Define more than one Model on a spécific table